requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
pyyaml